Bug 9859: fix argument type to PCRE function
When the flag is PCRE_INFO_OPTIONS, the pcre_fullinfo() expects
an argument of type unsigned long int.

+++ lib/rpcs_job/job.c
@@ -219,7 +219,7 @@ regexp_data_create(char *pattern, unsigned int extract,
                    regexp_data_t **regexp_data)
 {
     regexp_data_t *result = NULL;
-    unsigned int option_bits;
+    unsigned long int option_bits;
     const char *error;
     te_errno rc = 0;
     int erroffset;
